+++ Source/Modules/xISHServer/Install-ISHToolDotNET.ISH12.ps1
@@ -1,4 +1,5 @@
 function Install-ISHToolDotNET 
 {
-    Write-Warning "Assuming .NET 4.5 is installed on Windows Server 2012 R2"
+    $osInfo=Get-ISHOSInfo
+    Write-Warning "Assuming .NET 4.5 is installed on $($osInfo.Caption)"
 }