+++ Source/Scripts/PowerShellGet/Install-Module.ps1
@@ -1,13 +1,17 @@
 ﻿param (
     [Parameter(Mandatory=$false)]
     [string[]]$Computer,
+    [Parameter(Mandatory=$false)]
+    [pscredential]$Credential=$null,
     [Parameter(Mandatory=$true)]
     [string[]]$ModuleName,
     [Parameter(Mandatory=$false)]
     [string]$Repository,
     [Parameter(Mandatory=$false)]
     [ValidateSet("AllUsers","CurrentUser")]
-    [string]$Scope="AllUsers"
+    [string]$Scope="AllUsers",
+    [Parameter(Mandatory=$false)]
+    [switch]$AllowClobber=$false
 )    
 
 $cmdletsPaths="$PSScriptRoot\..\..\Cmdlets"
@@ -33,13 +37,19 @@ $installScriptBlock={
 
         if(-not $latestModule)
         {
-            Write-Error "Could not find package $name"
+            Write-Error "Could not find module $name"
             return
         }
         Write-Verbose "Found module $name with version $($latestModule.Version)"
-        Write-Verbose "Installing module $name with version $($latestPackage.Version)"
 
-        $latestModule|Install-Module -Scope $Scope -Force|Out-Null
+        if($AllowClobber)
+        {
+            $latestModule|Install-Module -Scope $Scope -AllowClobber -Force|Out-Null
+        }
+        else
+        {
+            $latestModule|Install-Module -Scope $Scope -Force|Out-Null
+        }
    
         Write-Host "Installed module $name with version $($latestModule.Version)"
     }
@@ -48,7 +58,7 @@ $installScriptBlock={
 #Install the packages
 try
 {
-    Invoke-CommandWrap -ComputerName $Computer -ScriptBlock $installScriptBlock -BlockName "Install Modules $ModuleName" -UseParameters @("ModuleName","Repository","Scope")
+    Invoke-CommandWrap -ComputerName $Computer -Credential $Credential -ScriptBlock $installScriptBlock -BlockName "Install Modules $ModuleName" -UseParameters @("ModuleName","Repository","Scope","AllowClobber")
 }
 catch
 {